demo functions are updated
The original `simple_demo_box()` function performed a demo in a three-
dimensional space. The fucnction has been renamed to `simple_demo_box3d()`.
The funtion that performs a demo in a two-dimensional box is appended as
`simple_demo_box()`.

The `seed` option is appended to both the two demo functions, which fixes
the state of the random generator.